requirements.txt
numpy>=1.24.0
networkx>=3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.0.0
xgboost>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

